requirements.txt
selenium~=4.14.0
pytest~=7.4.2
allure-pytest~=2.13.2